Move functionality to S3BlobContainer
