+++ plugins/repository-s3/src/main/java/org/opensearch/repositories/s3/utils/HttpRangeUtils.java
@@ -10,28 +10,26 @@ package org.opensearch.repositories.s3.utils;
 
 import software.amazon.awssdk.core.exception.SdkException;
 
-import org.opensearch.common.collect.Tuple;
-
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public final class HttpRangeUtils {
-    private static final Pattern RANGE_PATTERN = Pattern.compile("^bytes\\s+(\\d+)-(\\d+)/(\\d+|.*)$");
+    private static final Pattern RANGE_PATTERN = Pattern.compile("^bytes\\s+(\\d+)-\\d+[/\\d*]+$");
 
     /**
-     * Parses the content range header string value to calculate the start and end of the HTTP response.
+     * Parses the content range header string value to calculate the start (offset) of the HTTP response.
      * Tests against the RFC9110 specification of content range string.
      * Sample values: "bytes 0-10/200", "bytes 0-10/*"
      *  <a href="https://www.rfc-editor.org/rfc/rfc9110.html#name-content-range">Details here</a>
      * @param headerValue Header content range string value from the HTTP response
-     * @return Pair of values where v1 represents the lower and v2 represents the upper bound of the stream
+     * @return Start (Offset) value of the HTTP response
      */
-    public static Tuple<Long, Long> fromHttpRangeHeader(String headerValue) {
+    public static Long getStartOffsetFromRangeHeader(String headerValue) {
         Matcher matcher = RANGE_PATTERN.matcher(headerValue);
         if (!matcher.find()) {
             throw SdkException.create("Regex match for Content-Range header {" + headerValue + "} failed", new RuntimeException());
         }
-        return new Tuple<>(Long.parseLong(matcher.group(1)), Long.parseLong(matcher.group(2)));
+        return Long.parseLong(matcher.group(1));
     }
 
     /**
