+++ plugins/repository-s3/src/test/java/org/opensearch/repositories/s3/utils/HttpRangeUtilsTests.java
@@ -10,23 +10,20 @@ package org.opensearch.repositories.s3.utils;
 
 import software.amazon.awssdk.core.exception.SdkException;
 
-import org.opensearch.common.collect.Tuple;
 import org.opensearch.test.OpenSearchTestCase;
 
 public final class HttpRangeUtilsTests extends OpenSearchTestCase {
 
     public void testFromHttpRangeHeader() {
         String headerValue = "bytes 0-10/200";
-        Tuple<Long, Long> range = HttpRangeUtils.fromHttpRangeHeader(headerValue);
-        assertEquals(0L, range.v1().longValue());
-        assertEquals(10L, range.v2().longValue());
+        Long offset = HttpRangeUtils.getStartOffsetFromRangeHeader(headerValue);
+        assertEquals(0L, offset.longValue());
 
         headerValue = "bytes 0-10/*";
-        range = HttpRangeUtils.fromHttpRangeHeader(headerValue);
-        assertEquals(0L, range.v1().longValue());
-        assertEquals(10L, range.v2().longValue());
+        offset = HttpRangeUtils.getStartOffsetFromRangeHeader(headerValue);
+        assertEquals(0L, offset.longValue());
 
         final String invalidHeaderValue = "bytes */*";
-        assertThrows(SdkException.class, () -> HttpRangeUtils.fromHttpRangeHeader(invalidHeaderValue));
+        assertThrows(SdkException.class, () -> HttpRangeUtils.getStartOffsetFromRangeHeader(invalidHeaderValue));
     }
 }
